backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
SpeechRecognition==3.10.0
pydub==0.25.1
openai-whisper==20231117
torch==2.2.2+cu118 --index-url https://download.pytorch.org/whl/cu118
transformers==4.35.2
diffusers==0.24.0
psutil==5.9.6
python-dotenv==1.0.0
accelerate==0.24.1
safetensors>=0.4.0
Pillow>=10.0.1
numpy>=1.24.3
scipy>=1.11.4
requests>=2.31.0
huggingface-hub==0.19.4 